Set level to the single passed stream.
Fixes #10.
